data_consumer/requirements.txt
celery
redis
sqlalchemy-mixins==2.0.5
psycopg2-binary
pydantic==2.0.3
python-dotenv==0.19.2